+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Implement comments using Prettier's API.
+- Fix unstable comments in between If/Else blocks.
 
 ## 1.0.0-beta.2
 - Fix long static method calls producing undeployable code ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/69)).