+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix unstable comments for NestedExpr.
 - Fix unstable formatting for Method Declaration with no body.
 - Fix unstable comments for Annotations.
+- Fix unstable leading comments to ValueWhen and ElseWhen blocks.
 
 ## 1.0.0-beta.2
 - Fix long static method calls producing undeployable code ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/69)).